Consolidate more of TypeWrapper, remove jsg/value.h
