--- src/workerd/jsg/value.h
@@ -1,698 +1,0 @@
-// Copyright (c) 2017-2022 Cloudflare, Inc.
-// Licensed under the Apache 2.0 license found in the LICENSE file or at:
-//     https://opensource.org/licenses/Apache-2.0
-
-#pragma once
-// INTERNAL IMPLEMENTATION FILE
-//
-// Handling of various basic value types: numbers, booleans, strings, optionals, maybes, variants,
-// arrays, buffers, dicts.
-
-#include <workerd/jsg/fast-api.h>
-#include <workerd/jsg/util.h>
-#include <workerd/jsg/web-idl.h>
-#include <workerd/jsg/wrappable.h>
-
-#include <v8-container.h>
-#include <v8-date.h>
-
-#include <kj/debug.h>
-#include <kj/one-of.h>
-#include <kj/time.h>
-
-namespace workerd::jsg {
-
-// =======================================================================================
-// OneOf / variants
-
-template <typename T>
-constexpr bool isOneOf = false;
-template <typename... T>
-constexpr bool isOneOf<kj::OneOf<T...>> = true;
-// TODO(cleanup): Move to kj/one-of.h?
-
-// TypeWrapper mixin for variants.
-template <typename TypeWrapper>
-class OneOfWrapper {
- public:
-  template <typename... U>
-  static kj::String getName(kj::OneOf<U...>*) {
-    const auto getNameStr = [](auto u) {
-      if constexpr (kj::isSameType<const std::type_info&, decltype(TypeWrapper::getName(u))>()) {
-        return typeName(TypeWrapper::getName(u));
-      } else {
-        return kj::str(TypeWrapper::getName(u));
-      }
-    };
-
-    return kj::strArray(kj::arr(getNameStr(static_cast<U*>(nullptr))...), " or ");
-  }
-
-  template <typename U, typename... V>
-  bool wrapHelper(Lock& js,
-      v8::Local<v8::Context> context,
-      kj::Maybe<v8::Local<v8::Object>> creator,
-      kj::OneOf<V...>& in,
-      v8::Local<v8::Value>& out) {
-    if (in.template is<U>()) {
-      out =
-          static_cast<TypeWrapper*>(this)->wrap(js, context, creator, kj::mv(in.template get<U>()));
-      return true;
-    } else {
-      return false;
-    }
-  }
-
-  template <typename... U>
-  v8::Local<v8::Value> wrap(Lock& js,
-      v8::Local<v8::Context> context,
-      kj::Maybe<v8::Local<v8::Object>> creator,
-      kj::OneOf<U...> value) {
-    v8::Local<v8::Value> result;
-    if (!(wrapHelper<U>(js, context, creator, value, result) || ...)) {
-      result = js.undefined();
-    }
-    return result;
-  }
-
-  template <template <typename> class Predicate, typename U, typename... V>
-  bool unwrapHelperRecursive(
-      Lock& js, v8::Local<v8::Context> context, v8::Local<v8::Value> in, kj::OneOf<V...>& out) {
-    if constexpr (isOneOf<U>) {
-      // Ugh, a nested OneOf. We can't just call tryUnwrap(), because then our string/numeric
-      // coercion might trigger early.
-      U val;
-      if (unwrapHelper<Predicate>(js, context, in, val)) {
-        out.template init<U>(kj::mv(val));
-        return true;
-      }
-    } else if constexpr (Predicate<kj::Decay<U>>::value) {
-      KJ_IF_SOME(val,
-          static_cast<TypeWrapper*>(this)->tryUnwrap(js, context, in, (U*)nullptr, kj::none)) {
-        out.template init<U>(kj::mv(val));
-        return true;
-      }
-    }
-    return false;
-  }
-
-  template <template <typename> class Predicate, typename... U>
-  bool unwrapHelper(
-      Lock& js, v8::Local<v8::Context> context, v8::Local<v8::Value> in, kj::OneOf<U...>& out) {
-    return (unwrapHelperRecursive<Predicate, U>(js, context, in, out) || ...);
-  }
-
-  // Predicates for helping implement nested OneOf unwrapping. These must be struct templates
-  // because we can't pass variable templates as template template parameters.
-
-  template <typename T>
-  struct IsResourceType {
-    static constexpr bool value = webidl::isNonCallbackInterfaceType<T>;
-  };
-  template <typename T>
-  struct IsFallibleType {
-    static constexpr bool value =
-        !(webidl::isStringType<T> || webidl::isNumericType<T> || webidl::isBooleanType<T>);
-  };
-  template <typename T>
-  struct IsStringType {
-    static constexpr bool value = webidl::isStringType<T>;
-  };
-  template <typename T>
-  struct IsNumericType {
-    static constexpr bool value = webidl::isNumericType<T>;
-  };
-  template <typename T>
-  struct IsBooleanType {
-    static constexpr bool value = webidl::isBooleanType<T>;
-  };
-
-  template <typename... U>
-  kj::Maybe<kj::OneOf<U...>> tryUnwrap(Lock& js,
-      v8::Local<v8::Context> context,
-      v8::Local<v8::Value> handle,
-      kj::OneOf<U...>*,
-      kj::Maybe<v8::Local<v8::Object>> parentObject) {
-    (void)webidl::UnionTypeValidator<kj::OneOf<U...>>();
-    // Just need to instantiate this, static_asserts will do the rest.
-
-    // In order for string, numeric, and boolean coercion to function as expected, we need to follow
-    // the algorithm defined by Web IDL section 3.2.22 to convert JS values to OneOfs. That
-    // algorithm is written in a terribly wonky way, of course, but it appears we can restate it
-    // like so:
-    //
-    //   - Perform a series of breadth-first-searches on the OneOf, filtering out certain categories
-    //     of types on each run. For the types which are not filtered out, perform a tryUnwrap() on
-    //     that type, and succeed if that call succeeds (i.e., short-circuit). The filters used for
-    //     each pass are the following:
-    //       a. Consider only fallible (uncoercible) types.
-    //       b. If the JS value is a boolean, consider only boolean types.
-    //       c. If the JS value is a number, consider only numeric types.
-    //       d. Consider only string types.
-    //       e. Consider only numeric types.
-    //       f. Consider only boolean types.
-    //
-    // Note the symmetry across steps b-f. This way, strings only get coerced to numbers if the
-    // OneOf doesn't contain a string type, numbers only get coerced to strings if the OneOf doesn't
-    // contain a numeric type, objects only get coerced to a coercible type if there's no matching
-    // object type, null and undefined only get coerced to a coercible type if there's no nullable
-    // type, etc.
-    //
-    // TODO(soon): Hacked this by unwrapping into resource types first, so that we can unwrap
-    //   Requests and Responses into Initializers without them being interpreted as dictionaries. I
-    //   believe this is actually what the Web IDL spec prescribes anyway, but verify.
-    //
-    // TODO(someday): Prove that this is the same algorithm as the one defined by Web IDL.
-    kj::OneOf<U...> result;
-    if (unwrapHelper<IsResourceType>(js, context, handle, result) ||
-        unwrapHelper<IsFallibleType>(js, context, handle, result) ||
-        (handle->IsBoolean() && unwrapHelper<IsBooleanType>(js, context, handle, result)) ||
-        (handle->IsNumber() && unwrapHelper<IsNumericType>(js, context, handle, result)) ||
-        (handle->IsBigInt() && unwrapHelper<IsNumericType>(js, context, handle, result)) ||
-        (unwrapHelper<IsStringType>(js, context, handle, result)) ||
-        (unwrapHelper<IsNumericType>(js, context, handle, result)) ||
-        (unwrapHelper<IsBooleanType>(js, context, handle, result))) {
-      return kj::mv(result);
-    }
-    return kj::none;
-  }
-};
-
-// =======================================================================================
-// Arrays
-
-// TypeWrapper mixin for arrays.
-template <typename TypeWrapper>
-class ArrayWrapper {
- public:
-  static auto constexpr MAX_STACK = 64;
-  template <typename U>
-  static constexpr const char* getName(kj::Array<U>*) {
-    return "Array";
-  }
-
-  template <typename U>
-  v8::Local<v8::Value> wrap(Lock& js,
-      v8::Local<v8::Context> context,
-      kj::Maybe<v8::Local<v8::Object>> creator,
-      kj::Array<U> array) {
-    v8::Isolate* isolate = js.v8Isolate;
-    v8::EscapableHandleScope handleScope(isolate);
-
-    v8::LocalVector<v8::Value> items(isolate, array.size());
-    for (auto n = 0; n < items.size(); n++) {
-      items[n] = static_cast<TypeWrapper*>(this)->wrap(js, context, creator, kj::mv(array[n]));
-    }
-    auto out = v8::Array::New(isolate, items.data(), items.size());
-
-    return handleScope.Escape(out);
-  }
-  template <typename U>
-  v8::Local<v8::Value> wrap(Lock& js,
-      v8::Local<v8::Context> context,
-      kj::Maybe<v8::Local<v8::Object>> creator,
-      kj::ArrayPtr<U> array) {
-    v8::Isolate* isolate = js.v8Isolate;
-    v8::EscapableHandleScope handleScope(isolate);
-
-    v8::LocalVector<v8::Value> items(isolate, array.size());
-    for (auto n = 0; n < items.size(); n++) {
-      items[n] = static_cast<TypeWrapper*>(this)->wrap(js, context, creator, kj::mv(array[n]));
-    }
-    auto out = v8::Array::New(isolate, items.data(), items.size());
-
-    return handleScope.Escape(out);
-  }
-  template <typename U>
-  v8::Local<v8::Value> wrap(Lock& js,
-      v8::Local<v8::Context> context,
-      kj::Maybe<v8::Local<v8::Object>> creator,
-      kj::Array<U>& array) {
-    return static_cast<TypeWrapper*>(this)->wrap(js, context, creator, array.asPtr());
-  }
-
-  template <typename U>
-  kj::Maybe<kj::Array<U>> tryUnwrap(Lock& js,
-      v8::Local<v8::Context> context,
-      v8::Local<v8::Value> handle,
-      kj::Array<U>*,
-      kj::Maybe<v8::Local<v8::Object>> parentObject) {
-    if (!handle->IsArray()) {
-      return kj::none;
-    }
-
-    auto array = handle.As<v8::Array>();
-    auto length = array->Length();
-    auto builder = kj::heapArrayBuilder<U>(length);
-    for (auto i: kj::zeroTo(length)) {
-      v8::Local<v8::Value> element = check(array->Get(context, i));
-      builder.add(static_cast<TypeWrapper*>(this)->template unwrap<U>(
-          js, context, element, TypeErrorContext::arrayElement(i)));
-    }
-    return builder.finish();
-  }
-};
-
-// =======================================================================================
-// Sets
-
-// TypeWrapper mixin for sets.
-template <typename TypeWrapper>
-class SetWrapper {
- public:
-  static auto constexpr MAX_STACK = 64;
-  template <typename U>
-  static constexpr const char* getName(kj::HashSet<U>*) {
-    return "Set";
-  }
-
-  template <typename U>
-  v8::Local<v8::Value> wrap(Lock& js,
-      v8::Local<v8::Context> context,
-      kj::Maybe<v8::Local<v8::Object>> creator,
-      kj::HashSet<U> set) {
-    v8::Isolate* isolate = js.v8Isolate;
-    v8::EscapableHandleScope handleScope(isolate);
-
-    auto out = v8::Set::New(isolate);
-    for (auto& item: set) {
-      v8::HandleScope scope(isolate);
-      check(out->Add(
-          context, static_cast<TypeWrapper*>(this)->wrap(js, context, creator, kj::mv(item))));
-    }
-
-    return handleScope.Escape(out);
-  }
-
-  template <typename U>
-  kj::Maybe<kj::HashSet<U>> tryUnwrap(Lock& js,
-      v8::Local<v8::Context> context,
-      v8::Local<v8::Value> handle,
-      kj::HashSet<U>*,
-      kj::Maybe<v8::Local<v8::Object>> parentObject) {
-    if (!handle->IsSet()) {
-      return kj::none;
-    }
-
-    auto set = handle.As<v8::Set>();
-    auto array = set->AsArray();
-    auto length = array->Length();
-    auto builder = kj::HashSet<U>();
-    builder.reserve(length);
-    for (auto i: kj::zeroTo(length)) {
-      v8::Local<v8::Value> element = check(array->Get(context, i));
-      auto value = static_cast<TypeWrapper*>(this)->template unwrap<U>(
-          js, context, element, TypeErrorContext::other());
-      builder.upsert(kj::mv(value), [&](U& existing, U&& replacement) {
-        JSG_FAIL_REQUIRE(TypeError, "Duplicate values in the set after unwrapping.");
-      });
-    }
-    return kj::mv(builder);
-  }
-};
-
-// =======================================================================================
-// ArrayBuffers / ArrayBufferViews
-//
-// This wrapper implements the following wrapping conversions:
-//  - kj::Array<kj::byte> -> ArrayBuffer
-//
-// And the following unwrapping conversions:
-//   - ArrayBuffer -> kj::Array<kj::byte>
-//     (the kj::Array object holds a Global to the unwrapped ArrayBuffer)
-//   - ArrayBufferView -> kj::Array<kj::byte>
-//     (the kj::Array object holds a Global to the unwrapped ArrayBufferView's backing buffer)
-//
-// Note that there are no conversions for kj::ArrayPtr<kj::byte>, since it does not own its own
-// buffer -- fine in C++, but problematic in a GC language like JS. Restricting the interface to
-// only operate on owned arrays makes memory management simpler and safer in both directions.
-//
-// Logically a kj::Array<byte> could be considered analogous to a Uint8Array in JS, and for a time
-// that was the wrapping conversion implemented by this wrapper. However, the most common use cases
-// in web platform APIs involve accepting BufferSources for processing as immutable input and
-// returning ArrayBuffers. Since a kj::byte does not map to any JavaScript primitive, establishing
-// a mapping between ArrayBuffer/ArrayBufferView and Array<byte> is unambiguous and
-// convenient. The few places where a specific TypedArray is expected (e.g. Uint8Array) can be
-// handled explicitly with a v8::Local<v8::Uint8Array> (or other appropriate TypedArray type).
-//
-// BufferSource arguments to web platform API methods are typically expected to be processed but not
-// mutated, such as the input parameter to TextDecoder.decode(). This processing might happen
-// asynchronously, such as the plaintext parameter to SubtleCrypto.encrypt(). I am unaware of any
-// use of BufferSources which involve mutating the underlying ArrayBuffer -- typically an explicit
-// ArrayBufferView is expected for this case, such as the parameters to crypto.getRandomValues() or
-// the Streams spec's BYOB reader's read() method.
-//
-// This suggests the following rules of thumb:
-//
-// 1. If a BufferSource parameter is used as input to a:
-//   - synchronous method: accept a `kj::Array<const kj::byte>`.
-//   - asynchronous method (user is allowed to re-use the buffer during processing): accept a
-//     `kj::Array<const kj::byte>` and explicitly copy its bytes.
-//
-// 2. If a method accepts an ArrayBufferView that it is expected to mutate:
-//   - accept a `v8::Local<v8::ArrayBufferView>` explicitly (handled by V8HandleWrapper in
-//     type-wrapper.h) rather than a `kj::Array<kj::byte>` -- otherwise your method's contract
-//     will be wider than intended.
-//   - use `jsg::asBytes()` as a quick way to get a `kj::ArrayPtr<kj::byte>` view onto it.
-//
-// 3. If a method returns an ArrayBuffer, create and return a `kj::Array<kj::byte>`.
-template <typename TypeWrapper>
-class ArrayBufferWrapper {
- public:
-  static constexpr const char* getName(kj::ArrayPtr<byte>*) {
-    return "ArrayBuffer or ArrayBufferView";
-  }
-  static constexpr const char* getName(kj::ArrayPtr<const byte>*) {
-    return "ArrayBuffer or ArrayBufferView";
-  }
-  static constexpr const char* getName(kj::Array<byte>*) {
-    return "ArrayBuffer or ArrayBufferView";
-  }
-
-  v8::Local<v8::ArrayBuffer> wrap(jsg::Lock& js,
-      v8::Local<v8::Context> context,
-      kj::Maybe<v8::Local<v8::Object>> creator,
-      kj::Array<byte> value) {
-    return wrap(js.v8Isolate, creator, kj::mv(value));
-  }
-
-  v8::Local<v8::ArrayBuffer> wrap(
-      v8::Isolate* isolate, kj::Maybe<v8::Local<v8::Object>> creator, kj::Array<byte> value) {
-    // We need to construct a BackingStore that owns the byte array. We use the version of
-    // v8::ArrayBuffer::NewBackingStore() that accepts a deleter callback, and arrange for it to
-    // delete an Array<byte> placed on the heap.
-    //
-    size_t size = value.size();
-    if (size == 0) {
-      // BackingStore doesn't call custom deleter if begin is null, which it often is for empty
-      // arrays.
-      return v8::ArrayBuffer::New(isolate, 0);
-    }
-    byte* begin = value.begin();
-    if (isolate->GetGroup().SandboxContains(begin)) {
-      // TODO(perf): We could avoid an allocation here, perhaps, by decomposing the
-      //   kj::Array<byte> into its component pointer and disposer, and then pass the disposer
-      //   pointer as the "deleter_data" for NewBackingStore. However, KJ doesn't give us any way
-      //   to decompose an Array<T> this way, and it might not want to, as this could make it
-      //   impossible to support unifying Array<T> and Vector<T> in the future (i.e. making all
-      //   Array<T>s growable). So it may be best to stick with allocating an Array<byte> on the
-      //   heap after all...
-      auto ownerPtr = new kj::Array<byte>(kj::mv(value));
-
-      std::unique_ptr<v8::BackingStore> backing = v8::ArrayBuffer::NewBackingStore(
-          begin, size, [](void* begin, size_t size, void* ownerPtr) {
-        delete reinterpret_cast<kj::Array<byte>*>(ownerPtr);
-      }, ownerPtr);
-      KJ_REQUIRE(backing != nullptr, "Failed to create ArrayBuffer backing store");
-
-      return v8::ArrayBuffer::New(isolate, kj::mv(backing));
-    } else {
-      // The Array is not already inside the sandbox.  We have to make a copy and move it in.
-      // For performance reasons we might want to throw here and fix all callers to allocate
-      // inside the sandbox.
-      auto& js = Lock::from(isolate);
-      auto in_sandbox = js.allocBackingStore(size, Lock::AllocOption::UNINITIALIZED);
-
-      memcpy(in_sandbox->Data(), value.begin(), size);
-
-      return v8::ArrayBuffer::New(isolate, kj::mv(in_sandbox));
-    }
-  }
-
-  kj::Maybe<kj::Array<byte>> tryUnwrap(Lock& js,
-      v8::Local<v8::Context> context,
-      v8::Local<v8::Value> handle,
-      kj::Array<byte>*,
-      kj::Maybe<v8::Local<v8::Object>> parentObject) {
-    if (handle->IsArrayBufferView()) {
-      return asBytes(handle.As<v8::ArrayBufferView>());
-    } else if (handle->IsArrayBuffer()) {
-      return asBytes(handle.As<v8::ArrayBuffer>());
-    }
-    return kj::none;
-  }
-
-  kj::Maybe<kj::Array<const byte>> tryUnwrap(Lock& js,
-      v8::Local<v8::Context> context,
-      v8::Local<v8::Value> handle,
-      kj::Array<const byte>*,
-      kj::Maybe<v8::Local<v8::Object>> parentObject) {
-    return tryUnwrap(js, context, handle, static_cast<kj::Array<byte>*>(nullptr), parentObject);
-  }
-};
-
-// =======================================================================================
-// Dicts
-
-// TypeWrapper mixin for dictionaries (objects used as string -> value maps).
-template <typename TypeWrapper>
-class DictWrapper {
- public:
-  template <typename K, typename V>
-  static constexpr const char* getName(Dict<V, K>*) {
-    return "object";
-  }
-
-  template <typename K, typename V>
-  v8::Local<v8::Value> wrap(Lock& js,
-      v8::Local<v8::Context> context,
-      kj::Maybe<v8::Local<v8::Object>> creator,
-      Dict<V, K> dict) {
-    static_assert(webidl::isStringType<K>, "Dicts must be keyed on a string type.");
-
-    v8::Isolate* isolate = js.v8Isolate;
-    v8::EscapableHandleScope handleScope(isolate);
-    auto out = v8::Object::New(isolate);
-    for (auto& field: dict.fields) {
-      // Set() returns Maybe<bool>. As usual, if the Maybe is null, then there was an exception,
-      // but I have no idea what it means if the Maybe was filled in with the boolean value false...
-      KJ_ASSERT(check(out->Set(context,
-          static_cast<TypeWrapper*>(this)->wrap(js, context, creator, kj::mv(field.name)),
-          static_cast<TypeWrapper*>(this)->wrap(js, context, creator, kj::mv(field.value)))));
-    }
-    return handleScope.Escape(out);
-  }
-
-  template <typename K, typename V>
-  kj::Maybe<Dict<V, K>> tryUnwrap(Lock& js,
-      v8::Local<v8::Context> context,
-      v8::Local<v8::Value> handle,
-      Dict<V, K>*,
-      kj::Maybe<v8::Local<v8::Object>> parentObject) {
-    static_assert(webidl::isStringType<K>, "Dicts must be keyed on a string type.");
-
-    auto& wrapper = static_cast<TypeWrapper&>(*this);
-
-    // Currently the same as wrapper.unwrap<kj::String>(), but this allows us not to bother with the
-    // TypeErrorContext, or worrying about whether the tryUnwrap(kj::String*) version will ever be
-    // modified to return nullptr in the future.
-    const auto convertToUtf8 = [isolate = js.v8Isolate](v8::Local<v8::String> v8String) {
-      auto buf = kj::heapArray<char>(v8String->Utf8LengthV2(isolate) + 1);
-      v8String->WriteUtf8V2(
-          isolate, buf.begin(), buf.size(), v8::String::WriteFlags::kNullTerminate);
-      return kj::String(kj::mv(buf));
-    };
-
-    if (!handle->IsObject() || handle->IsArray()) {
-      return kj::none;
-    }
-
-    auto object = handle.As<v8::Object>();
-    v8::Local<v8::Array> names = check(object->GetOwnPropertyNames(context));
-    auto length = names->Length();
-    auto builder = kj::heapArrayBuilder<typename Dict<V, K>::Field>(length);
-    for (auto i: kj::zeroTo(length)) {
-      v8::Local<v8::String> name = check(check(names->Get(context, i))->ToString(context));
-      v8::Local<v8::Value> value = check(object->Get(context, name));
-
-      if constexpr (kj::isSameType<K, kj::String>()) {
-        auto strName = convertToUtf8(name);
-        const char* cstrName = strName.cStr();
-        builder.add(typename Dict<V, K>::Field{kj::mv(strName),
-          wrapper.template unwrap<V>(
-              js, context, value, TypeErrorContext::dictField(cstrName), object)});
-      } else {
-        // Here we have to be a bit more careful than for the kj::String case. The unwrap<K>() call
-        // may throw, but we need the name in UTF-8 for the very exception that it needs to throw.
-        // Thus, we do the unwrapping manually and UTF-8-convert the name only if it's needed.
-        auto unwrappedName = wrapper.tryUnwrap(js, context, name, static_cast<K*>(nullptr), object);
-        if (unwrappedName == kj::none) {
-          auto strName = convertToUtf8(name);
-          throwTypeError(js.v8Isolate, TypeErrorContext::dictKey(strName.cStr()),
-              TypeWrapper::getName(static_cast<K*>(nullptr)));
-        }
-        auto unwrappedValue =
-            wrapper.tryUnwrap(js, context, value, static_cast<V*>(nullptr), object);
-        if (unwrappedValue == kj::none) {
-          auto strName = convertToUtf8(name);
-          throwTypeError(js.v8Isolate, TypeErrorContext::dictField(strName.cStr()),
-              TypeWrapper::getName(static_cast<V*>(nullptr)));
-        }
-        builder.add(typename Dict<V, K>::Field{
-          KJ_ASSERT_NONNULL(kj::mv(unwrappedName)), KJ_ASSERT_NONNULL(kj::mv(unwrappedValue))});
-      }
-    }
-    return Dict<V, K>{builder.finish()};
-  }
-};
-
-// =======================================================================================
-// NonCoercible<T>
-
-template <typename TypeWrapper>
-class NonCoercibleWrapper {
- public:
-  template <CoercibleType T>
-  static auto getName(NonCoercible<T>*) {
-    return TypeWrapper::getName(static_cast<T*>(nullptr));
-  }
-
-  template <CoercibleType T>
-  v8::Local<v8::Value> wrap(Lock& js,
-      v8::Local<v8::Context> context,
-      kj::Maybe<v8::Local<v8::Object>> creator,
-      NonCoercible<T>) = delete;
-
-  template <CoercibleType T>
-  kj::Maybe<NonCoercible<T>> tryUnwrap(Lock& js,
-      v8::Local<v8::Context> context,
-      v8::Local<v8::Value> handle,
-      NonCoercible<T>*,
-      kj::Maybe<v8::Local<v8::Object>> parentObject) {
-    auto& wrapper = static_cast<TypeWrapper&>(*this);
-    if constexpr (kj::isSameType<kj::String, T>() || kj::isSameType<jsg::USVString, T>() ||
-        kj::isSameType<jsg::DOMString, T>()) {
-      if (!handle->IsString()) return kj::none;
-      KJ_IF_SOME(value,
-          wrapper.tryUnwrap(js, context, handle, static_cast<T*>(nullptr), parentObject)) {
-        return NonCoercible<T>{
-          .value = kj::mv(value),
-        };
-      }
-      return kj::none;
-    } else if constexpr (kj::isSameType<bool, T>()) {
-      if (!handle->IsBoolean()) return kj::none;
-      return wrapper.tryUnwrap(js, context, handle, static_cast<T*>(nullptr), parentObject)
-          .map([](auto& value) {
-        return NonCoercible<T>{
-          .value = value,
-        };
-      });
-    } else if constexpr (kj::isSameType<double, T>()) {
-      if (!handle->IsNumber()) return kj::none;
-      return wrapper.tryUnwrap(js, context, handle, static_cast<T*>(nullptr), parentObject)
-          .map([](auto& value) {
-        return NonCoercible<T>{
-          .value = value,
-        };
-      });
-    } else {
-      return nullptr;
-    }
-  }
-};
-
-// =======================================================================================
-// kj::Exception
-//
-// The kj::Exception wrapper handles the translation of so-called "tunneled" exceptions
-// between KJ and JavaScript. The wrapper is capable of turning any JavaScript value into
-// a kj::Exception with the caveat that the kj::Exception is not guaranteed to retain all
-// of the detail. Likewise, it can turn a kj::Exception with the correct metadata into a
-// reasonable JavaScript exception.
-
-class DOMException;
-
-template <typename TypeWrapper>
-class ExceptionWrapper {
- public:
-  static constexpr const char* getName(kj::Exception*) {
-    return "Exception";
-  }
-
-  v8::Local<v8::Value> wrap(Lock& js,
-      v8::Local<v8::Context> context,
-      kj::Maybe<v8::Local<v8::Object>> creator,
-      kj::Exception exception) {
-    return js.exceptionToJsValue(kj::mv(exception)).getHandle(js);
-  }
-
-  kj::Maybe<kj::Exception> tryUnwrap(Lock& js,
-      v8::Local<v8::Context> context,
-      v8::Local<v8::Value> handle,
-      kj::Exception*,
-      kj::Maybe<v8::Local<v8::Object>> parentObject) {
-
-    // If handle is a DOMException, then createTunneledException will not work
-    // here. We have to manually handle the DOMException case.
-    //
-    // Note that this is a general issue with any JSG_RESOURCE_TYPE that we
-    // happen to use as Errors. The createTunneledException() method uses V8's
-    // ToDetailString() to extract the detail about the error in a manner that
-    // is safe and side-effect free. Unfortunately, that mechanism does not
-    // work for JSG_RESOURCE_TYPE objects that are used as errors. For those,
-    // we need to drop down to the C++ interface and generate the kj::Exception
-    // ourselves. If any additional JSG_RESOURCE_TYPE error-like things are
-    // introduced, they'll need to be handled explicitly here also.
-    auto& wrapper = TypeWrapper::from(js.v8Isolate);
-    kj::Exception result = [&]() {
-      kj::Exception::Type excType = [&]() {
-        // Use .retryable and .overloaded properties as hints for what kj exception type to use.
-        if (handle->IsObject()) {
-          auto object = handle.As<v8::Object>();
-
-          if (js.toBool(check(object->Get(context, v8StrIntern(js.v8Isolate, "overloaded"_kj))))) {
-            return kj::Exception::Type::OVERLOADED;
-          }
-          if (js.toBool(check(object->Get(context, v8StrIntern(js.v8Isolate, "retryable"_kj))))) {
-            return kj::Exception::Type::DISCONNECTED;
-          }
-        }
-        return kj::Exception::Type::FAILED;
-      }();
-
-      KJ_IF_SOME(domException,
-          wrapper.tryUnwrap(js, context, handle, (DOMException*)nullptr, parentObject)) {
-        return KJ_EXCEPTION(FAILED,
-            kj::str("jsg.DOMException(", domException.getName(), "): ", domException.getMessage()));
-      } else {
-
-        static const constexpr kj::StringPtr PREFIXES[] = {
-#define V(name, _) name##_kj,
-          JS_ERROR_TYPES(V)
-#undef V
-              "DOMException"_kj,
-        };
-
-        kj::String reason;
-        if (!handle->IsObject()) {
-          // if the argument isn't an object, it couldn't possibly be an Error.
-          reason = kj::str(JSG_EXCEPTION(Error) ": ", handle);
-        } else {
-          reason = kj::str(handle);
-          bool found = false;
-          // If the error message starts with a platform error type that we tunnel,
-          // prefix it with "jsg."
-          for (auto name: PREFIXES) {
-            if (reason.startsWith(name)) {
-              reason = kj::str("jsg.", reason);
-              found = true;
-              break;
-            }
-          }
-          // Everything else should just come through as a normal error.
-          if (!found) {
-            reason = kj::str(JSG_EXCEPTION(Error) ": ", reason);
-          }
-        }
-        return kj::Exception(excType, __FILE__, __LINE__, kj::mv(reason));
-      }
-    }();
-
-    addExceptionDetail(js, result, handle);
-    addJsExceptionMetadata(js, result, handle);
-    return result;
-  }
-};
-
-}  // namespace workerd::jsg
